PMW-541 - resident management improvement

+++ app/Http/Controllers/ResidentDocumentController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\ResidentDocument\UpdateRequest;
 use App\Http\Resources\ResidentDocumentResource;
 use App\Http\Resources\ResidentDocumentResourceCollection;
 use App\Repositories\Contracts\ResidentDocumentRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 
 class ResidentDocumentController extends Controller
@@ -32,9 +33,12 @@ class ResidentDocumentController extends Controller
      *
      * @param IndexRequest $request
      * @return ResidentDocumentResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [ResidentDocument::class, $request->get('residentId')]);
+
         $residentDocuments = $this->residentDocumentRepository->findBy($request->all());
 
         return new ResidentDocumentResourceCollection($residentDocuments);
@@ -45,9 +49,12 @@ class ResidentDocumentController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @return ResidentDocumentResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [ResidentDocument::class, $request->get('residentId')]);
+
         $residentDocument = $this->residentDocumentRepository->save($request->all());
 
         return new ResidentDocumentResource($residentDocument);
@@ -58,9 +65,12 @@ class ResidentDocumentController extends Controller
      *
      * @param ResidentDocument $residentDocument
      * @return ResidentDocumentResource
+     * @throws AuthorizationException
      */
     public function show(ResidentDocument $residentDocument)
     {
+        $this->authorize('show', $residentDocument);
+
         return new ResidentDocumentResource($residentDocument);
     }
 
@@ -70,9 +80,12 @@ class ResidentDocumentController extends Controller
      * @param UpdateRequest $request
      * @param ResidentDocument $residentDocument
      * @return ResidentDocumentResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, ResidentDocument $residentDocument)
     {
+        $this->authorize('update', $residentDocument);
+
         $residentDocument = $this->residentDocumentRepository->update($residentDocument, $request->all());
 
         return new ResidentDocumentResource($residentDocument);
@@ -83,9 +96,12 @@ class ResidentDocumentController extends Controller
      *
      * @param ResidentDocument $residentDocument
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(ResidentDocument $residentDocument)
     {
+        $this->authorize('destroy', $residentDocument);
+
         $this->residentDocumentRepository->delete($residentDocument);
 
         return response()->json(null, 204);
